tiltfile: add stdin_mode='pty' to local_resouce()
Allows users to attach a pty,
for interactive commands that try
to grab the "main" tty

fixes https://github.com/tilt-dev/tilt/issues/6378
