+++ internal/tiltfile/docker.go
@@ -198,7 +198,7 @@ func (s *tiltfileState) dockerBuild(thread *starlark.Thread, fn *starlark.Builti
 		return nil, err
 	}
 
-	entrypointCmd, err := value.ValueToUnixCmd(thread, entrypoint, nil, nil)
+	entrypointCmd, err := value.ValueToUnixCmd(thread, entrypoint, nil, nil, value.StdinMode{})
 	if err != nil {
 		return nil, err
 	}
@@ -342,7 +342,7 @@ func (s *tiltfileState) customBuild(thread *starlark.Thread, fn *starlark.Builti
 		return nil, err
 	}
 
-	entrypointCmd, err := value.ValueToUnixCmd(thread, entrypoint, nil, nil)
+	entrypointCmd, err := value.ValueToUnixCmd(thread, entrypoint, nil, nil, value.StdinMode{})
 	if err != nil {
 		return nil, err
 	}
@@ -360,7 +360,8 @@ func (s *tiltfileState) customBuild(thread *starlark.Thread, fn *starlark.Builti
 		commandBat = commandBatVal
 	}
 
-	command, err := value.ValueGroupToCmdHelper(thread, commandVal, commandBat, dir, env)
+	command, err := value.ValueGroupToCmdHelper(thread,
+		commandVal, commandBat, dir, env, value.StdinMode{})
 	if err != nil {
 		return nil, fmt.Errorf("Argument 2 (command): %v", err)
 	} else if command.Empty() {
