+++ internal/tiltfile/files.go
@@ -55,7 +55,9 @@ func (s *tiltfileState) local(thread *starlark.Thread, fn *starlark.Builtin, arg
 		return nil, err
 	}
 
-	cmd, err := value.ValueGroupToCmdHelper(thread, commandValue, commandBatValue, commandDirValue, commandEnv)
+	cmd, err := value.ValueGroupToCmdHelper(thread,
+		commandValue, commandBatValue, commandDirValue,
+		commandEnv, value.StdinMode{})
 	if err != nil {
 		return nil, err
 	}
