+++ internal/tiltfile/local_resource.go
@@ -52,6 +52,7 @@ func (s *tiltfileState) localResource(thread *starlark.Thread, fn *starlark.Buil
 	var allowParallel bool
 	var links links.LinkList
 	var labels value.LabelSet
+	var updateStdinMode, serveStdinMode value.StdinMode
 	autoInit := true
 	if fn.Name() == testN {
 		// If we're initializing a test, by default parallelism is on
@@ -78,6 +79,8 @@ func (s *tiltfileState) localResource(thread *starlark.Thread, fn *starlark.Buil
 		"readiness_probe?", &readinessProbe,
 		"dir?", &updateCmdDirVal,
 		"serve_dir?", &serveCmdDirVal,
+		"stdin_mode??", &updateStdinMode,
+		"serve_stdin_mode??", &serveStdinMode,
 	); err != nil {
 		return nil, err
 	}
@@ -92,11 +95,15 @@ func (s *tiltfileState) localResource(thread *starlark.Thread, fn *starlark.Buil
 		return nil, err
 	}
 
-	updateCmd, err := value.ValueGroupToCmdHelper(thread, updateCmdVal, updateCmdBatVal, updateCmdDirVal, updateEnv)
+	updateCmd, err := value.ValueGroupToCmdHelper(thread,
+		updateCmdVal, updateCmdBatVal, updateCmdDirVal, updateEnv,
+		updateStdinMode)
 	if err != nil {
 		return nil, err
 	}
-	serveCmd, err := value.ValueGroupToCmdHelper(thread, serveCmdVal, serveCmdBatVal, serveCmdDirVal, serveEnv)
+	serveCmd, err := value.ValueGroupToCmdHelper(thread,
+		serveCmdVal, serveCmdBatVal, serveCmdDirVal, serveEnv,
+		serveStdinMode)
 	if err != nil {
 		return nil, err
 	}
