+++ internal/tiltfile/telemetry/telemetry.go
@@ -39,7 +39,7 @@ func setTelemetryCmd(thread *starlark.Thread, fn *starlark.Builtin, args starlar
 		return starlark.None, err
 	}
 
-	cmd, err := value.ValueGroupToCmdHelper(thread, cmdVal, cmdBatVal, cmdDirVal, nil)
+	cmd, err := value.ValueGroupToCmdHelper(thread, cmdVal, cmdBatVal, cmdDirVal, nil, value.StdinMode{})
 	if err != nil {
 		return nil, err
 	}
